# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  lead.AssignLeadRequest:
    properties:
      admin_id:
        type: string
      priority:
        type: integer
    required:
    - admin_id
    type: object
  lead.ConvertLeadRequest:
    properties:
      bin:
        type: string
      legal_address:
        type: string
      legal_name:
        type: string
      org_type:
        enum:
        - ip
        - llp
        type: string
      password:
        minLength: 8
        type: string
    required:
    - bin
    - legal_address
    - legal_name
    - org_type
    - password
    type: object
  lead.LeadListResponse:
    properties:
      leads:
        items:
          $ref: '#/definitions/lead.OwnerLead'
        type: array
      total:
        type: integer
    type: object
  lead.OwnerLead:
    properties:
      assigned_to:
        type: string
      bin:
        type: string
      company_name:
        description: Company info
        type: string
      contact_email:
        type: string
      contact_name:
        description: Contact person
        type: string
      contact_phone:
        type: string
      contact_position:
        type: string
      converted_at:
        description: Conversion
        type: string
      converted_user_id:
        type: integer
      created_at:
        type: string
      follow_up_count:
        type: integer
      how_found_us:
        type: string
      id:
        type: integer
      last_contacted_at:
        description: Follow-up
        type: string
      legal_address:
        type: string
      next_follow_up_at:
        type: string
      notes:
        type: string
      priority:
        type: integer
      referrer_url:
        type: string
      rejection_reason:
        type: string
      source:
        description: UTM tracking
        type: string
      status:
        allOf:
        - $ref: '#/definitions/lead.Status'
        description: Lead management
      updated_at:
        type: string
      use_case:
        description: Application details
        type: string
      utm_campaign:
        type: string
      utm_medium:
        type: string
      utm_source:
        type: string
      website:
        type: string
    type: object
  lead.Status:
    enum:
    - new
    - contacted
    - qualified
    - converted
    - rejected
    - lost
    type: string
    x-enum-varnames:
    - StatusNew
    - StatusContacted
    - StatusQualified
    - StatusConverted
    - StatusRejected
    - StatusLost
  lead.SubmitLeadRequest:
    properties:
      bin:
        type: string
      company_name:
        description: Company info
        type: string
      contact_email:
        type: string
      contact_name:
        description: Contact person
        type: string
      contact_phone:
        type: string
      contact_position:
        type: string
      how_found_us:
        type: string
      legal_address:
        type: string
      use_case:
        description: Application details
        type: string
      utm_campaign:
        type: string
      utm_medium:
        type: string
      utm_source:
        description: UTM parameters (optional, from query params)
        type: string
      website:
        type: string
    required:
    - company_name
    - contact_email
    - contact_name
    - contact_phone
    type: object
  lead.UpdateLeadStatusRequest:
    properties:
      notes:
        type: string
      reason:
        description: For rejection
        type: string
      status:
        allOf:
        - $ref: '#/definitions/lead.Status'
        enum:
        - new
        - contacted
        - qualified
        - rejected
        - lost
    required:
    - status
    type: object
  lead.swaggerConvertLeadResponse:
    properties:
      email:
        type: string
      message:
        type: string
      success:
        type: boolean
      user_id:
        type: integer
    type: object
  lead.swaggerLeadListResponse:
    properties:
      data:
        $ref: '#/definitions/lead.LeadListResponse'
      success:
        type: boolean
    type: object
  lead.swaggerLeadResponse:
    properties:
      data:
        $ref: '#/definitions/lead.OwnerLead'
      success:
        type: boolean
    type: object
  lead.swaggerLeadStatsResponse:
    properties:
      data: {}
      success:
        type: boolean
    type: object
...
  response.ErrorData:
    properties:
      code:
        type: string
      details: {}
      error_trace:
        description: Full error + stack trace (non-prod only)
        type: string
      message:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/response.ErrorData'
      success:
        example: false
        type: boolean
    type: object
  response.SuccessResponse:
    properties:
      message:
        example: Success
        type: string
      success:
        example: true
        type: boolean
    type: object
...
info:
  contact: {}
  description: API server for booking system.
  title: PhotoStudio API
  version: "1.0"
paths:
...
  /admin/leads:
    get:
      description: Админский эндпоинт для просмотра заявок с возможностью фильтрации
        по статусу.
      parameters:
      - description: Фильтр по статусу (new, contacted, qualified, converted, rejected)
        in: query
        name: status
        type: string
      - description: 'Лимит (дефолт: 50)'
        in: query
        name: limit
        type: integer
      - description: Отступ
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список заявок
          schema:
            $ref: '#/definitions/lead.swaggerLeadListResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Список лидов
      tags:
      - Admin Leads
  /admin/leads/{id}:
    get:
      description: Админский эндпоинт для получения заявки по её идентификатору.
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Данные заявки
          schema:
            $ref: '#/definitions/lead.swaggerLeadResponse'
        "400":
          description: Некорректный ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Заявка не найдена
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить лид
      tags:
      - Admin Leads
  /admin/leads/{id}/assign:
    patch:
      consumes:
      - application/json
      description: Админ назначает заявку на определенного менеджера и задает приоритет.
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для назначения
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/lead.AssignLeadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Менеджер назначен
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Ошибка входных данных
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Заявка не найдена
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "422":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Назначить лида
      tags:
      - Admin Leads
  /admin/leads/{id}/contacted:
    post:
      description: Меняет статус заявки на contacted.
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Заявка отмечена как Contacted
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Некорректный ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Заявка не найдена
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Отметить лида как Contacted
      tags:
      - Admin Leads
  /admin/leads/{id}/convert:
    post:
      consumes:
      - application/json
      description: Создает пользователя-владельца (owner) из заявки (статус converted)
        и профиль компании.
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: Данные конвертации (Юр. адрес и тд)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/lead.ConvertLeadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Лид успешно конвертирован в пользователя
          schema:
            $ref: '#/definitions/lead.swaggerConvertLeadResponse'
        "400":
          description: Нельзя конвертировать (например, некорректный статус) или неверный
            запрос
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Заявка не найдена
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Email уже зарегистрирован
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "422":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Конвертировать лида
      tags:
      - Admin Leads
  /admin/leads/{id}/reject:
    post:
      consumes:
      - application/json
      description: Админ отклоняет заявку (статус rejected) с указанием причины.
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: Укажите reason (причину отказа)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/lead.UpdateLeadStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Заявка отклонена
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Заявка уже конвертирована или неверный ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Заявка не найдена
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Отклонить лида
      tags:
      - Admin Leads
  /admin/leads/{id}/status:
    patch:
      consumes:
      - application/json
      description: Админский эндпоинт. Позволяет вручную указать статус, примечания
        и причину отказа.
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: Новый статус
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/lead.UpdateLeadStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Статус обновлен
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Ошибка входных данных
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Заявка не найдена
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Заявка уже конвертирована
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "422":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновить статус лида
      tags:
      - Admin Leads
  /admin/leads/stats:
    get:
      description: Админский эндпоинт для просмотра воронки заявок (счетчики по статусам).
      produces:
      - application/json
      responses:
        "200":
          description: Успех
          schema:
            $ref: '#/definitions/lead.swaggerLeadStatsResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Статистика по лидам
      tags:
      - Admin Leads
...
  /leads/submit:
    post:
      consumes:
      - application/json
      description: Публичный эндпоинт для отправки заявки на регистрацию студии.
      parameters:
      - description: Данные заявки
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/lead.SubmitLeadRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Заявка создана
          schema:
            $ref: '#/definitions/lead.swaggerLeadResponse'
        "400":
          description: Ошибка входных данных
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Email уже существует
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "422":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Подать заявку лида
      tags:
      - Leads
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"